# python-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rebyu BKT Service",
    "description": "Bayesian Knowledge Tracing microservice for Rebyu. It trains pyBKT models, stores lesson parameters, and updates learner mastery in real time.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/bkt/mastery/events": {
      "post": {
        "tags": [
          "mastery"
        ],
        "summary": "Ingest Mastery Event",
        "operationId": "ingest_mastery_event_api_v1_bkt_mastery_events_post",
        "parameters": [
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasteryEventCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasteryEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bkt/mastery/events/batch": {
      "post": {
        "tags": [
          "mastery"
        ],
        "summary": "Ingest Mastery Events",
        "operationId": "ingest_mastery_events_api_v1_bkt_mastery_events_batch_post",
        "parameters": [
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasteryEventBatchCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasteryEventResponse"
                  },
                  "title": "Response Ingest Mastery Events Api V1 Bkt Mastery Events Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bkt/mastery/learners/{learner_id}": {
      "get": {
        "tags": [
          "mastery"
        ],
        "summary": "Get Learner Mastery",
        "operationId": "get_learner_mastery_api_v1_bkt_mastery_learners__learner_id__get",
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          },
          {
            "name": "lesson_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lesson Id"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerMasteryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bkt/mastery/learners/{learner_id}/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "mastery"
        ],
        "summary": "Get Lesson Mastery",
        "operationId": "get_lesson_mastery_api_v1_bkt_mastery_learners__learner_id__lessons__lesson_id__get",
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerLessonMasteryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mastery"
        ],
        "summary": "Reset Lesson Mastery",
        "operationId": "reset_lesson_mastery_api_v1_bkt_mastery_learners__learner_id__lessons__lesson_id__delete",
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          },
          {
            "name": "delete_event_history",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Delete Event History"
            }
          },
          {
            "name": "x-service-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LearnerLessonMasteryResponse": {
        "properties": {
          "learner_id": {
            "type": "integer",
            "title": "Learner Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "mastery_probability": {
            "type": "number",
            "title": "Mastery Probability"
          },
          "mastery_level": {
            "type": "string",
            "title": "Mastery Level"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "last_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Id"
          },
          "model_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Run Id"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "lesson_id",
          "mastery_probability",
          "mastery_level",
          "attempt_count",
          "last_event_id",
          "model_run_id",
          "last_updated"
        ],
        "title": "LearnerLessonMasteryResponse"
      },
      "LearnerMasteryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LearnerLessonMasteryResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "average_mastery_probability": {
            "type": "number",
            "title": "Average Mastery Probability"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "average_mastery_probability"
        ],
        "title": "LearnerMasteryListResponse"
      },
...
      "MasteryEventBatchCreate": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/MasteryEventCreate"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "MasteryEventBatchCreate"
      },
      "MasteryEventCreate": {
        "properties": {
          "source_event_id": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Source Event Id"
          },
          "learner_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Learner Id"
          },
          "lesson_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Lesson Id"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "difficulty_level": {
            "type": "string",
            "enum": [
              "EASY",
              "AVERAGE",
              "HARD"
            ],
            "title": "Difficulty Level"
          },
          "assessment_type": {
            "type": "string",
            "enum": [
              "DIAGNOSTIC",
              "LESSON_QUIZ",
              "MIDDLE_EXAM",
              "MOCK_EXAM"
            ],
            "title": "Assessment Type"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "source_event_id",
          "learner_id",
          "lesson_id",
          "is_correct",
          "difficulty_level",
          "assessment_type"
        ],
        "title": "MasteryEventCreate"
      },
      "MasteryEventResponse": {
        "properties": {
          "source_event_id": {
            "type": "string",
            "title": "Source Event Id"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          },
          "learner_id": {
            "type": "integer",
            "title": "Learner Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "predicted_correct_probability": {
            "type": "number",
            "title": "Predicted Correct Probability"
          },
          "mastery_before": {
            "type": "number",
            "title": "Mastery Before"
          },
          "mastery_posterior": {
            "type": "number",
            "title": "Mastery Posterior"
          },
          "mastery_after": {
            "type": "number",
            "title": "Mastery After"
          },
          "mastery_level": {
            "type": "string",
            "title": "Mastery Level"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "parameters_used": {
            "$ref": "#/components/schemas/ParametersUsed"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "source_event_id",
          "duplicate",
          "learner_id",
          "lesson_id",
          "question_id",
          "is_correct",
          "predicted_correct_probability",
          "mastery_before",
          "mastery_posterior",
          "mastery_after",
          "mastery_level",
          "attempt_count",
          "parameters_used",
          "processed_at"
        ],
        "title": "MasteryEventResponse"
      },
...
      "ParametersUsed": {
        "properties": {
          "prior": {
            "type": "number",
            "title": "Prior"
          },
          "learn": {
            "type": "number",
            "title": "Learn"
          },
          "guess": {
            "type": "number",
            "title": "Guess"
          },
          "slip": {
            "type": "number",
            "title": "Slip"
          },
          "forget": {
            "type": "number",
            "title": "Forget"
          },
          "model_variant": {
            "type": "string",
            "title": "Model Variant"
          }
        },
        "type": "object",
        "required": [
          "prior",
          "learn",
          "guess",
          "slip",
          "forget",
          "model_variant"
        ],
        "title": "ParametersUsed"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}